requirements.txt
altgraph==0.17.5
anyio==4.12.1
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
comtypes==1.4.15; sys_platform == 'win32'
exceptiongroup==1.3.1
googletrans==4.0.2
gTTS==2.5.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
mtranslate==1.8
packaging==26.0
pefile==2024.8.26; sys_platform == 'win32'
pillow==12.1.0
playsound==1.2.2
psutil==7.2.1
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pypiwin32==223; sys_platform == 'win32'
PyGObject==3.46.0; sys_platform == 'linux'
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.18.0
pyqtkeybind==0.0.9
pytesseract==0.3.13
pyttsx3==2.99
pywin32==311; sys_platform == 'win32'
pywin32-ctypes==0.2.3; sys_platform == 'win32'
requests==2.32.5
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
